feat(detect): clone the correct pgaccess_test() fucntion, and change the pgacess() funciton prototype to be suitable with the pgaccess_test()

+++ user/pgtbltest.c
@@ -143,30 +143,24 @@ superpg_test()
   printf("superpg_test: OK\n");  
 }
 
+
 void
 pgaccess_test()
 {
   char *buf;
-  uint64 abits;
-
+  unsigned int abits;
   printf("pgaccess_test starting\n");
   testname = "pgaccess_test";
-
   buf = malloc(32 * PGSIZE);
-
   if (pgaccess(buf, 32, &abits) < 0)
     err("pgaccess failed");
-
   buf[PGSIZE * 1] += 1;
   buf[PGSIZE * 2] += 1;
   buf[PGSIZE * 30] += 1;
-
   if (pgaccess(buf, 32, &abits) < 0)
     err("pgaccess failed");
-
-  if (abits != ((1ULL << 1) | (1ULL << 2) | (1ULL << 30)))
+  if (abits != ((1 << 1) | (1 << 2) | (1 << 30)))
     err("incorrect access bits set");
-
   free(buf);
   printf("pgaccess_test: OK\n");
 }
